requirements.txt
omegaconf
hydra-core
pytorch
pytorch-cuda=12.1
torchaudio
pysoundfile
sentencepiece
pybind11
folly
eigen
moviepy
